Created EpisodesControls component that houses the navigation controls for Episodes. Took render functionality from Episodes renderUtils.
