+++ src/components/Episodes/renderUtils.jsx
@@ -1,19 +1,6 @@
 import React from "react";
 import Episode from "COMPONENTS/Episode/";
 
-function renderPaginationLink(data){
-
-	const { number } = data;
-
-	return (
-		<li key={`episode__pagination__${number}`}>
-			<a href={ `#episode-${number}` }>
-				{ number }
-			</a>
-		</li>
-	);	
-}// renderPaginationLink
-
 function renderEpisode(data, index){
 
 	const { 
@@ -38,6 +25,5 @@ function renderEpisode(data, index){
 }// renderEpisode
 
 export default { 
-	pagination: renderPaginationLink,
 	episode: renderEpisode
 };
