+++ src/components/EpisodesControls/index.js
@@ -1,0 +1,4 @@
+import RENDER from "./renderUtils.jsx";
+
+export { RENDER };
+export { default } from "./EpisodesControls.jsx";
